Reduce visibility of "edit on Github" button
When you visit a page in the documentation right now, the object on the page immediately attracting the most attention is "Edit on Github", which strikes me as not really the most important thing on the page compared to the likely actual documentation the user is seeking. This edit just reduces its visibility by making it grey rather than the red it was set to.