Implement ability to have external sidebar category links.
Previously, only actual pages within the docs were permitted to be
linked from the _config.yml `sidebar_categories` array.  This builds on
the existing behavior, but allows an object to be inserted in place of
the normal element with the `href` and `title` attributes:

```yaml
sidebar_categories:
  Usage:
    - downloading-schema
    - mutations
    - href: http://excite.com/
      title: Excite.com
```

It's worth pointing out that this won't work with the "find adjacent"
logic in place within `page.ejs` which enables the Next and Previous
navigation buttons on individual pages.  In the event that an external
link is met, the next/previous link simply will not show.

It's possible this can be improved to at least show the button with an
external destination, though it's arguable if sending the user to
another domain is an appreciable behavior.