Copy directory into place, rather than using symlinks.
It seems that Netlify might not be following symlinks properly when
looking at the "directory to publish".  Probably to avoid cycles, and
maybe due to the default configuration on Go which doesn't follow
symlinks for this exact reason.

For a consistent experience, we'll just do this on Windows too.

Note that this isn't entirely atomic, as we don't move it along-side the
directory first and then rename it.