Remove the tools for looking up and extending a YAML config.
I had previously thought this was necessary in order to expand the
"root" (on top of it's existing value), but it's seeming like we can just set
the root to what we want (in the case of this repo, just the name of the
config we're testing).  While this could affect our ability to validate
failures on invalid 'root' values, we already were unable to do that as
Netlify is proxied through Fly.io in (most) of our actual deployments.