Add support for docsearch `appId` for custom search deployments.
This only seems to be important to me for the moment because of my need to
test search improvements.  Typically, Algolia docsearch runs their own free
deployment for open-source projects.  This is done through a public
configuration repository available at
https://github.com/algolia/docsearch-configs.

However, in order to test these configurations, it's necessary to run your
own deployment through their paid service (or a free trial!).  Their public
service defaults to their own app, but when using the paid service, it's
necessary to specify the `appId`, which this new code allows.