Upgrade `smooth-scroll` to latest version and switch to jsDelivr CDN.
The version of `smooth-scroll` we were using was quite old (v7.1.1) compared
to what is now available (v12).  By upgrading to the latest version, and
slightly tweaking the way we instantiate it to support the new construct,
this resolves the issue with anchor tags with periods being incorrectly
escaped.

Fixes https://github.com/meteor/hexo-theme-meteor/issues/57.