Separate each `script` tag into its own context to limit error scope.
Unfortunately, a result of the problem reported in #65 is that all other
JavaScript, including search, was prevented from working because they all
lived in a shared `<script>` tag.  This change puts each concern in its own
`<script>` tag to isolate the damage by those script tags.

Of course, each of the script providers would have had this in their own
code-snippets they offer for copy-pasting, but I guess we decided to put
them all in a single `<script>` tag at one point. :cry:

Ref: https://github.com/meteor/meteor-theme-hexo/commits/544f1f4084
Ref: https://github.com/meteor/meteor-theme-hexo/pull/65